Joost world parser

+++ src/gamelogic/state/Menu.cpp
@@ -70,7 +70,7 @@ namespace State {
         }
 
         this->_context.SDL_facade.draw_rect(
-            CoordinateDouble{120, this->get_selected()->coordinates.y + 6}, //todo delete this magic number
+            CoordinateDouble{(double)120, (double)this->get_selected()->coordinates.y + 6}, //todo delete this magic number
             15,
             15,
             _color
